+++ CullFactory/Plugin.cs
@@ -14,7 +14,7 @@ namespace CullFactory
 
         private const string Guid    = "com.fumiko.CullFactory";
         private const string Name    = "CullFactory";
-        private const string Version = "0.5.3";
+        private const string Version = "0.6.0";
 
         private void Awake()
         {